requirements.txt
tensorflow==2.8.0
numpy==1.22.3
Pillow==9.0.1
tensorboard==2.8.0
tensorboardX==2.4
matplotlib==3.5.1
scipy==1.8.0